População foi alterada para unsigned long int

+++ CartasSuperTrunfo.c
@@ -8,7 +8,7 @@ int main(){
     char estado1 = 'A';
     char codigo1[10] = "A01";
     char cidade1[20] = "São Paulo";
-    int populacao1 = 12325000;
+    unsigned long int populacao1 = 12325000;
     float area1 = 1521.11;
     float pib1 = 699.28;
     int pontos1 = 50;
@@ -19,7 +19,7 @@ int main(){
     printf("Estado: %c\n", estado1);
     printf("Código: %s\n", codigo1);
     printf("Nome da Cidade: %s\n", cidade1);
-    printf("População: %d\n", populacao1);
+    printf("População: %lu\n", populacao1);
     printf("Area: %.2f km²\n", area1);
     printf("PIB: %.2f bilhões de reais\n", pib1);
     printf("Números de Pontos Turísticos: %d\n", pontos1);
@@ -29,7 +29,7 @@ int main(){
     char estado2 = 'B';
     char codigo2[10] = "B02";
     char cidade2[20] = "Rio de Janeiro";
-    int populacao2 = 6748000;
+    unsigned long int populacao2 = 6748000;
     float area2 = 1200.25;
     float pib2 = 300.50;
     int pontos2 = 30;
@@ -40,7 +40,7 @@ int main(){
     printf("Estado: %c\n", estado2);
     printf("Código: %s\n", codigo2);
     printf("Nome da Cidade: %s\n", cidade2);
-    printf("População: %d\n", populacao2);
+    printf("População: %lu\n", populacao2);
     printf("Area: %.2f km²\n", area2);
     printf("PIB: %.2f bilhões de reais\n", pib2);
     printf("Números de Pontos Turísticos: %d\n", pontos2);
